+++ inlong-dashboard/src/configs/pagination/conf.ts
@@ -17,8 +17,15 @@
  * under the License.
  */
 
-import paginationConf from './conf';
+import i18n from '@/i18n';
 
-export const defaultSize = +paginationConf.pageSizeOptions[0];
+const showTotal = (total: number) => i18n.t('configs.pagination.Total', { total });
+
+const paginationConf = {
+  showQuickJumper: true,
+  showSizeChanger: true,
+  pageSizeOptions: ['10', '20', '50', '100'],
+  showTotal,
+};
 
 export default paginationConf;
